backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
alembic==1.13.3
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==11.0.0
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0
bcrypt==4.0.1
pdf2image==1.16.3
PyPDF2==3.0.1
